fix: Cleanup the deserialization code for analysis.
